refactor: use asm params in subprotocols (#1362)
* refactor: use admin roles from asm-params instead of primitives

* refactor: use admin params from asm-params

* refactor: use u16 for recovery delay

* refactor: use bridge params from asm-params

* refactor: use checkpoint params from asm-params

+++ crates/asm/spec/src/lib.rs
@@ -8,8 +8,9 @@ use std::num::NonZero;
 
 use bitcoin::secp256k1::{Parity, PublicKey};
 use strata_asm_common::{AsmSpec, Loader, Stage};
-use strata_asm_proto_administration::{AdministrationSubprotoParams, AdministrationSubprotocol};
-use strata_asm_proto_bridge_v1::{BridgeV1Config, BridgeV1Subproto};
+use strata_asm_params::{AdministrationSubprotoParams, BridgeV1Config};
+use strata_asm_proto_administration::AdministrationSubprotocol;
+use strata_asm_proto_bridge_v1::BridgeV1Subproto;
 use strata_asm_proto_checkpoint_v0::{
     CheckpointV0Params, CheckpointV0Subproto, CheckpointV0VerificationParams,
 };
@@ -82,7 +83,7 @@ impl StrataAsmSpec {
         let bridge_v1_genesis = BridgeV1Config {
             operators,
             denomination: BitcoinAmount::from_sat(params.deposit_amount.to_sat()),
-            assignment_duration: params.dispatch_assignment_dur as u64,
+            assignment_duration: params.dispatch_assignment_dur,
             // TODO(QQ): adjust
             operator_fee: BitcoinAmount::ZERO,
             recovery_delay: params.recovery_delay,
